pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = ""
authors = [
    {name = "dlsptm"}
]
packages = [
    { include = "app" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic (==1.14.0)",
    "annotated-types (==0.7.0)",
    "anyio (==4.7.0)",
    "bcrypt (==4.2.1)",
    "blinker (==1.9.0)",
    "cffi (==1.17.1)",
    "click (==8.1.8)",
    "contourpy (==1.3.1)",
    "cycler (==0.12.1)",
    "et-xmlfile (==2.0.0)",
    "fastapi (==0.115.6)",
    "flask (>=3.1.0,<4.0.0)",
    "flask-bcrypt (==1.0.1)",
    "flask-cors (==5.0.0)",
    "flask-login (>=0.6.3,<0.7.0)",
    "flask-migrate (==4.0.7)",
    "flask-sqlalchemy (==3.1.1)",
    "fonttools (==4.55.3)",
    "h11 (==0.14.0)",
    "idna (==3.10)",
    "itsdangerous (==2.2.0)",
    "jinja2 (==3.1.5)",
    "kiwisolver (==1.4.7)",
    "mako (==1.3.8)",
    "markupsafe (==3.0.2)",
    "matplotlib (==3.10.0)",
    "numpy (==2.2.1)",
    "openpyxl (==3.1.5)",
    "packaging (==24.2)",
    "pandas (==2.2.3)",
    "pillow (==11.0.0)",
    "psycopg2 (==2.9.10)",
    "pycparser (==2.22)",
    "pydantic (==2.10.4)",
    "pydantic-core (==2.27.2)",
    "pyjwt (==2.10.1)",
    "pyparsing (==3.2.0)",
    "python-dateutil (==2.9.0.post0)",
    "python-dotenv (==1.0.1)",
    "pytz (==2024.2)",
    "setuptools (==75.6.0)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "sqlalchemy (==2.0.36)",
    "starlette (==0.41.3)",
    "tornado (==6.4.2)",
    "typing-extensions (==4.12.2)",
    "tzdata (==2024.2)",
    "uvicorn (==0.34.0)",
    "werkzeug (>=3.1.3,<4.0.0)",
    "wheel (==0.45.1)",
    "networks (>=0.3.7,<0.4.0)",
    "pytest (>=8.3.5,<9.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"